Create hooks for each store in order details
